+++ backend/graphql/event/resolverEventShow.js
@@ -3,6 +3,7 @@ if (process.env.NODE_ENV !== 'production') {
 }
 const { UserInputError, AuthenticationError } = require('apollo-server')
 const Event = require('../../models/event')
+const EventRecurrence = require('../../models/eventRecurrence')
 const EventShow = require('../../models/eventShow')
 const EventEntry = require('../../models/eventEntry')
 
@@ -23,33 +24,43 @@ module.exports = {
 
               let eventToUpdate = await Event.findOne({ _id: args.eventId, owner: userId })
 
-              const newShow = new EventShow({
-                showname: args.showname,
-                description: args.description,
-                link: args.link,
-                duration: args.duration,
-                event: args.eventId,
-                recurrence: args.recurrenceId
-              })
-              const savedShow = await newShow.save()
+              if (eventToUpdate && eventToUpdate !== null) {
 
-              let newShows = eventToUpdate.shows.concat(savedShow)
+                let recurrenceToUpdate = await EventRecurrence.findOne({ event: args.eventId, _id: args.recurrenceId })
 
-              let updatedEvent = await Event.findOneAndUpdate(
-                { _id: args.eventId, owner: userId },
-                {
-                  $set: {
-                    shows: newShows
-                  }
-                },
-                {
-                  new: true
-                }
-              )
+                if (recurrenceToUpdate && recurrenceToUpdate !== null) {
+
+                  const newShow = new EventShow({
+                    showname: args.showname,
+                    description: args.description,
+                    link: args.link,
+                    duration: args.duration,
+                    event: args.eventId,
+                    recurrence: args.recurrenceId
+                  })
+                  const savedShow = await newShow.save()
+
+                  let newShows = recurrenceToUpdate.shows.concat(savedShow)
+
+                  let updatedRecurrence = await EventRecurrence.findOneAndUpdate(
+                    { _id: args.recurrenceId },
+                    {
+                      $set: {
+                        shows: newShows
+                      }
+                    },
+                    {
+                      new: true
+                    }
+                  )
 
               //console.log('Updated event is', updatedEvent)
 
-              return savedShow
+                  return savedShow
+
+                }
+              }
+
             } catch (error) {
               throw new UserInputError(error.message, {
                 invalidArgs: args,
@@ -74,8 +85,9 @@ module.exports = {
 
               let showToUpdate = await EventShow.findOne({ _id: args.id })
               let eventToUpdate = await Event.findOne({ _id: showToUpdate.event, owner: userId })
+              let recurrenceToUpdate = await EventRecurrence.findOne({ recurrence: showToUpdate.recurrence })
 
-              if (eventToUpdate && showToUpdate) {
+              if (eventToUpdate && showToUpdate && recurrenceToUpdate) {
 
                 showToUpdate.showname = args.showname
                 showToUpdate.description = args.description
@@ -84,12 +96,12 @@ module.exports = {
 
                 const updatedShow = await showToUpdate.save()
 
-                const newShows = eventToUpdate.shows.map((show) => {
+                const newShows = recurrenceToUpdate.shows.map((show) => {
                   return (show.id === args.id ? updatedShow : show)
                 })
 
-                let updatedEvent = await Event.findOneAndUpdate(
-                  { _id: eventToUpdate._id, owner: userId },
+                let updatedRecurrence = await EventRecurrence.findOneAndUpdate(
+                  { _id: recurrenceToUpdate._id },
                   {
                     $set: {
                       shows: newShows
@@ -99,7 +111,6 @@ module.exports = {
                     new: true
                   }
                 )
-                //console.log('Updated event is', updatedEvent)
 
                 return updatedShow
 
@@ -108,6 +119,8 @@ module.exports = {
                 throw new AuthenticationError('The event is not yours to update')
 
               }
+
+              return null
 
             } catch (error) {
               throw new UserInputError(error.message, {
@@ -142,8 +155,9 @@ module.exports = {
 
               let showToDelete = await EventShow.findOne({ _id: args.id })
               let eventToUpdate = await Event.findOne({ _id: showToDelete.event, owner: userId })
+              let recurrenceToUpdate = await EventRecurrence.findOne({ _id: showToDelete.recurrence })
 
-              if (eventToUpdate && showToDelete) {
+              if (eventToUpdate && recurrenceToUpdate && showToDelete) {
 
                 const result = await EventShow.deleteOne(
                   { _id: args.id },
@@ -151,10 +165,14 @@ module.exports = {
 
                 //console.log('Delete count is', result.deletedCount)
 
-                const newShows = eventToUpdate.shows.filter(show => show.id !== args.id)
+                if (result.deletedCount === 0) {
+                  return 0
+                }
 
-                let updatedEvent = await Event.findOneAndUpdate(
-                  { _id: eventToUpdate._id, owner: userId },
+                const newShows = recurrenceToUpdate.shows.filter(show => show.id !== args.id)
+
+                let updatedRecurrence = await EventRecurrence.findOneAndUpdate(
+                  { _id: recurrenceToUpdate._id },
                   {
                     $set: {
                       shows: newShows
@@ -164,7 +182,6 @@ module.exports = {
                     new: true
                   }
                 )
-                //console.log('Updated event is', updatedEvent)
 
                 return result.deletedCount
